Refactioring repository names

+++ src/main/java/com/ufape/agiota/negocio/service/AgiotaService.java
@@ -1,9 +1,11 @@
 package com.ufape.agiota.negocio.service;
 
+
+import org.springframework.security.access.AccessDeniedException;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
-import com.ufape.agiota.dados.repository.RepositoryAgiota;
+import com.ufape.agiota.dados.repository.AgiotaRepository;
 import com.ufape.agiota.negocio.models.Agiota;
 import java.util.List;
 
@@ -11,10 +13,12 @@ import lombok.RequiredArgsConstructor;
 
 @Service @RequiredArgsConstructor
 public class AgiotaService implements AgiotaServiceInterface{
-    final private RepositoryAgiota agiotaRepository;
+    final private AgiotaRepository agiotaRepository;
 
+    @Transactional
     @Override
-    public void delete(Long id) {
+    public void delete(Long id, String idSession) {
+        if (!find(id).getIdKc().equals(idSession)) throw new AccessDeniedException("User not allowed");
         agiotaRepository.deleteById(id);
     }
 
@@ -34,9 +38,11 @@ public class AgiotaService implements AgiotaServiceInterface{
         return agiotaRepository.findAll();
     }
 
+    @Transactional
     @Override
-    public Agiota update(Long id, Agiota agiota) {
-        return null;
+    public Agiota update(Agiota agiota, String idSession) {
+        if (!agiota.getIdKc().equals(idSession)) throw new AccessDeniedException("User not allowed");
+        return agiotaRepository.save(agiota);
     }
